requirements.txt
scipy==1.15.3
pypdf==6.3.0
google-genai==1.52.0
Flask==3.1.2
flask-cors==6.0.1
statsmodels==0.14.5
pandas==2.2.3
numpy==1.26.4
reportlab==4.2.5
python-dotenv==1.0.1
PyJWT==2.10.1